Renamed files to 'EmotiBit_', changed includes in files, changed version in library.properties
