+++ EmotiBit_Si7013.cpp
@@ -19,7 +19,7 @@ License along with this library; if not, write to the Free Software
 Foundation, Inc., 51 Franklin St, Fifth Floor, Boston, MA  02110-1301  USA
 */
 
-#include "CFL_Si7013.h"
+#include "EmotiBit_Si7013.h"
 
 Si7013::Si7013() {
   _humidityNew = false;
