+++ examples/EmotiBit_Si7013_Example/EmotiBit_Si7013_Example.ino
@@ -18,7 +18,7 @@ License along with this library; if not, write to the Free Software
 Foundation, Inc., 51 Franklin St, Fifth Floor, Boston, MA  02110-1301  USA
 */
 
-#include "CFL_Si7013.h"
+#include "EmotiBit_Si7013.h"
 
 #define SerialUSB SERIAL_PORT_USBVIRTUAL // Required to work in Visual Micro / Visual Studio IDE
 #define PRINT_SI7013_DATA 1
