+++ library.properties
@@ -1,5 +1,5 @@
 name=EmotiBit Si7013
-version=1.0.0
+version=0.0.1
 author=Connected Future Labs
 maintainer=Connected Future Labs <info@connectedfuturelabs.com>
 sentence=A library written for EmotiBit that supports the Si7013 temperature and humidity module
